Add state and animation Holding, do collison with koopa and enemy

+++ GameMario/game.cpp
@@ -72,7 +72,7 @@ void Game::Draw(float x, float y, LPDIRECT3DTEXTURE9 texture, int left, int top,
 	Camera* camera = Camera::GetInstance();
 	float cam_x, cam_y;
 	camera->GetCamPos(cam_x, cam_y);
-	D3DXVECTOR3 p(x - cam_x, y - cam_y, 0);
+	D3DXVECTOR3 p((int) (x - cam_x),(int) (y - cam_y), 0);
 	RECT r;
 	r.left = left;
 	r.top = top;
